requirements.txt
certifi==2023.5.7
charset-normalizer==3.1.0
idna==3.4
lxml==4.9.2
olefile==0.46
psycopg2==2.9.6
pypdf2==3.0.1
python-docx==0.8.11
python-dotenv==1.0.0
requests==2.30.0
typing-extensions==4.5.0
urllib3==2.0.2
pandas
beautifulsoup4
